+++ src/main/java/at/hannibal2/skyhanni/events/EntityHealthUpdateEvent.kt
@@ -1,0 +1,5 @@
+package at.hannibal2.skyhanni.events
+
+import net.minecraft.entity.EntityLivingBase
+
+class EntityHealthUpdateEvent(val entity: EntityLivingBase, val health: Float) : LorenzEvent()