+++ src/main/java/at/hannibal2/skyhanni/events/ResetEntityHurtEvent.kt
@@ -3,6 +3,6 @@ package at.hannibal2.skyhanni.events
 import net.minecraft.entity.EntityLivingBase
 import java.awt.Color
 
-class ResetEntityHurtTimeEvent(val entity: EntityLivingBase, var shouldReset: Boolean) : LorenzEvent()
+class ResetEntityHurtEvent(val entity: EntityLivingBase, var shouldReset: Boolean) : LorenzEvent()
 
 fun Color.withAlpha(alpha: Int): Int = (alpha.coerceIn(0, 255) shl 24) or (this.rgb and 0x00ffffff)