+++ src/main/java/at/hannibal2/skyhanni/features/chat/PlayerChatFilter.kt
@@ -7,15 +7,18 @@ import at.hannibal2.skyhanni.utils.LorenzLogger
 import at.hannibal2.skyhanni.utils.LorenzUtils
 import at.hannibal2.skyhanni.utils.StringUtils.removeColor
 import net.minecraftforge.fml.common.eventhandler.SubscribeEvent
+import java.util.regex.Pattern
 
 class PlayerChatFilter {
 
     private val loggerPlayerChat = LorenzLogger("chat/player")
 
+    //§8[§9109§8] §b[MVP§c+§b] 4Apex§f§r§f: omg selling
+    private val patternSkyBlockLevel = Pattern.compile("§8\\[§(.)(\\d+)§8] (.+)")
+
     @SubscribeEvent
     fun onChatMessage(event: LorenzChatEvent) {
         if (!LorenzUtils.isOnHypixel) return
-        if (!SkyHanniMod.feature.chat.playerMessages) return
 
         if (shouldBlock(event.message)) {
             event.blockedReason = "player_chat"
@@ -23,10 +26,19 @@ class PlayerChatFilter {
     }
 
     private fun shouldBlock(originalMessage: String): Boolean {
-        val split = if (originalMessage.contains("§7§r§7: ")) {
-            originalMessage.split("§7§r§7: ")
-        } else if (originalMessage.contains("§f: ")) {
-            originalMessage.split("§f: ")
+        //since hypixel sends own chat messages really weird " §r§8[§r§d205§r§8] §r§6[MVP§r§c++§r§6] hannibal2"
+        var rawMessage = originalMessage.replace("§r", "").trim()
+
+        val matcher = patternSkyBlockLevel.matcher(rawMessage)
+        if (matcher.matches()) {
+            SkyBlockLevelChatMessage.setData(matcher.group(2).toInt(), matcher.group(1))
+            rawMessage = matcher.group(3)
+        }
+
+        val split = if (rawMessage.contains("§7§7: ")) {
+            rawMessage.split("§7§7: ")
+        } else if (rawMessage.contains("§f: ")) {
+            rawMessage.split("§f: ")
         } else {
             return false
         }
@@ -36,6 +48,14 @@ class PlayerChatFilter {
 
         rawName = rawName.substring(channel.originalPrefix.length)
         val name = grabName(rawName) ?: return false
+
+        if (!SkyHanniMod.feature.chat.playerMessagesFormat) {
+            if (SkyHanniMod.feature.chat.hideSkyblockLevel) {
+                LorenzUtils.chat(rawMessage)
+                return true
+            }
+            return false
+        }
 
         val message = split[1]
         send(channel, name.removeColor(), message.removeColor())