+++ src/main/java/at/hannibal2/skyhanni/features/dungeon/DungeonCleanEnd.kt
@@ -1,17 +1,13 @@
 package at.hannibal2.skyhanni.features.dungeon
 
 import at.hannibal2.skyhanni.SkyHanniMod
-import at.hannibal2.skyhanni.events.CheckRenderEntityEvent
-import at.hannibal2.skyhanni.events.DamageIndicatorFinalBossEvent
-import at.hannibal2.skyhanni.events.LorenzChatEvent
-import at.hannibal2.skyhanni.events.PacketEvent
+import at.hannibal2.skyhanni.events.*
 import at.hannibal2.skyhanni.utils.LorenzUtils
 import at.hannibal2.skyhanni.utils.LorenzUtils.matchRegex
 import net.minecraft.client.Minecraft
 import net.minecraft.client.entity.EntityOtherPlayerMP
 import net.minecraft.entity.item.EntityArmorStand
 import net.minecraft.entity.monster.EntityGuardian
-import net.minecraft.network.play.server.S1CPacketEntityMetadata
 import net.minecraft.network.play.server.S2APacketParticles
 import net.minecraftforge.event.world.WorldEvent
 import net.minecraftforge.fml.common.eventhandler.SubscribeEvent
@@ -61,26 +57,17 @@ class DungeonCleanEnd {
     }
 
     @SubscribeEvent
-    fun onHealthUpdatePacket(event: PacketEvent.ReceiveEvent) {
+    fun onHealthUpdateEvent(event: EntityHealthUpdateEvent) {
         if (!LorenzUtils.inDungeons) return
         if (!SkyHanniMod.feature.dungeon.cleanEndToggle) return
-
         if (bossDone) return
         if (lastBossId == -1) return
+        if (event.entity.entityId != lastBossId) return
 
-        val packet = event.packet
-        if (packet !is S1CPacketEntityMetadata) return
-        if (packet.entityId != lastBossId) return
-
-        for (watchableObject in packet.func_149376_c()) {
-            if (watchableObject.dataValueId == 6) {
-                val health = watchableObject.`object` as Float
-                if (health < 1) {
-                    val dungeonFloor = DungeonData.dungeonFloor
-                    LorenzUtils.chat("§eFloor $dungeonFloor done!")
-                    bossDone = true
-                }
-            }
+        if (event.health < 1) {
+            val dungeonFloor = DungeonData.dungeonFloor
+            LorenzUtils.chat("§eFloor $dungeonFloor done!")
+            bossDone = true
         }
     }